Live code examples use .tsx with imports instead untyped .jsx
