+++ packages/website/docs/components/avatar/avatar.example.tsx
@@ -1,4 +1,6 @@
-function Page() {
+import { Avatar } from '@synergycodes/axiom';
+
+export function Example() {
   return (
     <Avatar
       imageUrl="https://picsum.photos/id/237/200/300"
