--- packages/website/docs/components/checkbox/checkbox.example.jsx
@@ -1,13 +1,0 @@
-function Page() {
-  return (
-    <div style={{ display: 'flex', gap: '1rem', alignItems: 'center' }}>
-      <Checkbox size="small" />
-      <Checkbox size="medium" />
-      <Checkbox size="extra-small" />
-      <Checkbox indeterminate />
-      <Checkbox disabled />
-      <Checkbox disabled checked />
-      <Checkbox disabled indeterminate />
-    </div>
-  );
-}
