--- packages/website/docs/components/menu/menu.example.jsx
@@ -1,7 +1,0 @@
-function Page() {
-  return (
-    <Menu items={[{ label: 'Star Wars' }, { label: 'Star Treck' }]}>
-      <LabelButton label="Open" />
-    </Menu>
-  );
-}
