+++ packages/website/docs/components/separator/separator-docs.tsx
@@ -1,6 +1,6 @@
 import { Separator } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./separator.example.jsx';
+import exampleCode from '!!raw-loader!./separator.example.tsx';
 
 export function SeparatorDocs() {
   return (
