--- packages/website/docs/components/snackbar/snackbar.example.jsx
@@ -1,22 +1,0 @@
-function Page() {
-  const [open, setOpen] = React.useState(false);
-
-  return (
-    <div>
-      {!open && (
-        <LabelButton onClick={() => setOpen(true)} label="Show Snackbar" />
-      )}
-      {open && (
-        <Snackbar
-          variant="success"
-          title="Operation completed successfully"
-          subtitle="Your changes have been saved"
-          close={true}
-          onClose={() => setOpen(false)}
-          onButtonClick={() => alert('Button clicked')}
-          buttonLabel="Undo"
-        />
-      )}
-    </div>
-  );
-}
