+++ packages/website/docs/components/switch/switch-docs.tsx
@@ -1,6 +1,6 @@
 import { Switch } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./switch.example.jsx';
+import exampleCode from '!!raw-loader!./switch.example.tsx';
 
 export function SwitchDocs() {
   return (
