+++ packages/website/docs/components/switch/switch.example.tsx
@@ -1,4 +1,6 @@
-function Page() {
+import { Switch } from "@synergycodes/axiom";
+
+export function Example() {
   return (
     <Switch
       size="medium"
