+++ packages/website/docs/components/tooltip/tooltip-content/tooltip-content-docs.tsx
@@ -1,6 +1,6 @@
 import { Tooltip, TooltipContent, TooltipTrigger } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./tooltip-content.example.jsx';
+import exampleCode from '!!raw-loader!./tooltip-content.example.tsx';
 
 export function TooltipContentDocs() {
   return (
