+++ packages/website/docs/components/tooltip/tooltip-content/tooltip-content.example.tsx
@@ -1,4 +1,6 @@
-function Page() {
+import { Tooltip, TooltipContent, TooltipTrigger } from '@synergycodes/axiom';
+
+export function Example() {
   return (
     <Tooltip open={true}>
       <TooltipTrigger>
