+++ packages/website/docs/components/tooltip/tooltip-docs.tsx
@@ -1,6 +1,6 @@
 import { Tooltip, TooltipContent, TooltipTrigger } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./tooltip.example.jsx';
+import exampleCode from '!!raw-loader!./tooltip.example.tsx';
 
 export function TooltipDocs() {
   return (
