+++ packages/website/docs/components/tooltip/tooltip-trigger/tooltip-trigger-docs.tsx
@@ -1,6 +1,6 @@
 import { Tooltip, TooltipTrigger } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./tooltip-trigger.example.jsx';
+import exampleCode from '!!raw-loader!./tooltip-trigger.example.tsx';
 
 export function TooltipTriggerDocs() {
   return (
