--- packages/website/docs/components/tooltip/tooltip.example.jsx
@@ -1,12 +1,0 @@
-function Page() {
-  return (
-    <Tooltip>
-      <TooltipTrigger>
-        <span>Tooltip</span>
-      </TooltipTrigger>
-      <TooltipContent>
-        <span>Tooltip</span>
-      </TooltipContent>
-    </Tooltip>
-  );
-}
