--- packages/website/tools/templates/jsx.txt
@@ -1,7 +1,0 @@
-function Page() {
-  return (
-    <componentPascalName
-        // TODO: add props
-    />
-  );
-}
