+++ packages/website/tools/templates/tsx.txt
@@ -1,6 +1,6 @@
 import { componentPascalName } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./ComponentKebabName.example.jsx';
+import exampleCode from '!!raw-loader!./ComponentKebabName.example.tsx';
 
 export function componentPascalNameDocs() {
   return (
@@ -10,7 +10,7 @@ export function componentPascalNameDocs() {
             // TODO: add props
         />
       }
-      // TODO: Check that the correct .css files are picked
+      // TODO: Check if the correct .css files are picked
       cssPaths={['components/ComponentKebabName/ComponentKebabName.module.css']}
       componentPath="components/ComponentKebabName/ComponentKebabName.tsx"
       exampleCode={exampleCode}
