+++ packages/website/tools/templates/tsxExample.txt
@@ -1,0 +1,9 @@
+import { componentPascalName } from "@synergycodes/axiom";
+
+export function Example() {
+  return (
+    <componentPascalName
+        // TODO: add props
+    />
+  );
+}
